cstl_lib.c

+++ src/local.c
@@ -147,7 +147,7 @@ static struct listener_t *current_listener;
 
 void do_alloc_uv_buffer(size_t suggested_size, uv_buf_t *buf) {
     char *tmp = (char *) malloc(suggested_size * sizeof(char));
-    *buf = uv_buf_init(tmp, suggested_size);
+    *buf = uv_buf_init(tmp, (unsigned int)suggested_size);
 }
 
 void do_dealloc_uv_buffer(uv_buf_t *buf) {
